+++ src/gleam_stdlib.erl
@@ -8,10 +8,10 @@
          bit_string_int_to_u32/1, bit_string_int_from_u32/1, decode_result/1,
          bit_string_slice/3, decode_bit_string/1, compile_regex/2, regex_scan/2,
          percent_encode/1, percent_decode/1, regex_check/2, regex_split/2,
-         base_decode64/1, parse_query/1, bit_string_concat/1, size_of_tuple/1, 
+         base_decode64/1, parse_query/1, bit_string_concat/1, size_of_tuple/1,
          decode_tuple/1, decode_tuple2/1, decode_tuple3/1, decode_tuple4/1,
-         decode_tuple5/1, decode_tuple6/1, tuple_get/2, classify_dynamic/1, 
-         print/1, println/1, print_error/1, println_error/1, inspect/1, 
+         decode_tuple5/1, decode_tuple6/1, tuple_get/2, classify_dynamic/1,
+         print/1, println/1, print_error/1, println_error/1, inspect/1,
          float_to_string/1, int_from_base_string/2]).
 
 %% Taken from OTP's uri_string module
@@ -26,6 +26,12 @@
            ((X) >= $A) andalso ((X) =< $F) -> (X) - $A + 10;
            ((X) >= $a) andalso ((X) =< $f) -> (X) - $a + 10
         end).
+
+-define(is_lowercase_char(X), (X > 96 andalso X < 123)).
+
+-define(is_underscore_char(X), (X == 95)).
+
+-define(is_digit_char(X), (X > 47 andalso X < 58)).
 
 map_get(Map, Key) ->
     case maps:find(Key, Map) of
@@ -363,14 +369,23 @@ inspect(true) ->
     "True";
 inspect(false) ->
     "False";
+inspect(nil) ->
+    "Nil";
 inspect(Any) when is_atom(Any) ->
-    lists:map(
-        fun(Part) ->
-            [Head | Tail] = string:next_grapheme(unicode:characters_to_binary(Part)),
-            [string:uppercase([Head]), Tail]
-        end,
-        re:split(erlang:atom_to_list(Any), "_+", [{return, iodata}])
-    );
+    AtomAsList = erlang:atom_to_list(Any),
+    case inspect_maybe_gleam_atom(AtomAsList, none, []) of
+        {ok, GleamCompatibleAtomString} -> erlang:list_to_binary(GleamCompatibleAtomString);
+        {error, Reason} -> case Reason of
+            % These atoms invalid in Gleam are quoted in Erlang:
+            cannot_be_an_empty_string -> ["//erl('')"];
+            cannot_start_with_a_digit -> ["//erl('", erlang:atom_to_binary(Any), "')"];
+            cannot_start_with_an_underscore -> ["//erl('", erlang:atom_to_binary(Any), "')"];
+            need_to_only_contain_lower_case_letters_aor_digits_aor_underscores -> ["//erl('", erlang:atom_to_binary(Any), "')"];
+            % These atoms invalid in Gleam are not quoted in Erlang:
+            cannot_contain_consecutive_underscores -> ["//erl(", erlang:atom_to_binary(Any), ")"];
+            cannot_end_with_an_underscore -> ["//erl(", erlang:atom_to_binary(Any), ")"]
+        end
+    end;
 inspect(Any) when is_integer(Any) ->
     erlang:integer_to_list(Any);
 inspect(Any) when is_float(Any) ->
@@ -410,6 +425,38 @@ inspect(Any) when is_function(Any) ->
     ["//fn(", Args, ") { ... }"];
 inspect(Any) ->
     ["//erl(", io_lib:format("~p", [Any]), ")"].
+
+inspect_maybe_gleam_atom([], none, []) ->
+    {error, cannot_be_an_empty_string};
+inspect_maybe_gleam_atom([Head | _Rest], none, []) when ?is_digit_char(Head) ->
+    {error, cannot_start_with_a_digit};
+inspect_maybe_gleam_atom([$_ | _Rest], none, []) ->
+    {error, cannot_start_with_an_underscore};
+inspect_maybe_gleam_atom([$_ | []], _PrevChar, _Acc) ->
+    {error, cannot_end_with_an_underscore};
+inspect_maybe_gleam_atom([$_ | _Rest], $_, _Acc) ->
+    {error, cannot_contain_consecutive_underscores};
+inspect_maybe_gleam_atom([Head | _Rest], _PrevChar, _Acc)
+    when ?is_lowercase_char(Head) == false andalso ?is_underscore_char(Head) == false andalso ?is_digit_char(Head) == false ->
+    {error, need_to_only_contain_lower_case_letters_aor_digits_aor_underscores};
+% Handle first char -> uppercase and prepend
+inspect_maybe_gleam_atom([Head | Rest], none, Acc) ->
+    inspect_maybe_gleam_atom(Rest, Head, [string:uppercase([Head]) | Acc]);
+% Handle underscore -> only set as PrevChar
+inspect_maybe_gleam_atom([$_ | Rest], _PrevChar, Acc) ->
+    inspect_maybe_gleam_atom(Rest, $_, Acc);
+% Handle char after underscore -> uppercase and prepend
+inspect_maybe_gleam_atom([Head | Rest], $_, Acc) ->
+    inspect_maybe_gleam_atom(Rest, Head, [string:uppercase([Head]) | Acc]);
+% Handle char after number -> uppercase and prepend
+inspect_maybe_gleam_atom([Head | Rest], PrevChar, Acc) when ?is_digit_char(PrevChar) ->
+    inspect_maybe_gleam_atom(Rest, Head, [string:uppercase([Head]) | Acc]);
+% Handle any other char -> prepend only
+inspect_maybe_gleam_atom([Head | Rest], _PrevChar, Acc) ->
+    inspect_maybe_gleam_atom(Rest, Head, [Head | Acc]);
+% Handle end of string -> return reversed acc ok tuple
+inspect_maybe_gleam_atom([], _PrevChar, Acc) ->
+    {ok, lists:reverse(Acc)}.
 
 inspect_list([]) ->
     {proper, []};